+++ tests/interactive/text/test_caret_color.py
@@ -1,4 +1,6 @@
 import pytest
+
+import pyglet.text.layout.incremental
 from tests.base.interactive import InteractiveTestCase
 
 from pyglet import app
@@ -31,7 +33,7 @@ class TestWindow(window.Window):
         )
 
         self.margin = 2
-        self.layout = layout.IncrementalTextLayout(
+        self.layout = pyglet.text.layout.incremental.IncrementalTextLayout(
             self.document,
             self.width - self.margin * 2, self.height - self.margin * 2,
             multiline=True,
