+++ tests/interactive/text/test_content_valign.py
@@ -1,4 +1,6 @@
 import pytest
+
+import pyglet.text.layout.incremental
 from tests.base.interactive import InteractiveTestCase
 
 from pyglet import app
@@ -30,10 +32,10 @@ class TestWindow(window.Window):
         self.batch = graphics.Batch()
         self.document = text.decode_text(doctext)
         self.margin = 2
-        self.layout = layout.IncrementalTextLayout(self.document,
-            self.width - self.margin * 2, self.height - self.margin * 2,
-            multiline=True,
-            batch=self.batch)
+        self.layout = pyglet.text.layout.incremental.IncrementalTextLayout(self.document,
+                                                                           self.width - self.margin * 2, self.height - self.margin * 2,
+                                                                           multiline=True,
+                                                                           batch=self.batch)
         self.layout.content_valign = content_valign
         self.caret = caret.Caret(self.layout)
         self.push_handlers(self.caret)
