+++ tests/interactive/text/test_html.py
@@ -1,4 +1,6 @@
 import pytest
+
+import pyglet.text.layout.incremental
 from tests.base.interactive import InteractiveTestCase
 
 import pyglet
@@ -196,10 +198,10 @@ class TestWindow(pyglet.window.Window):
         self.batch = pyglet.graphics.Batch()
         self.document = pyglet.text.decode_html(doctext)
         self.margin = 2
-        self.layout = layout.IncrementalTextLayout(self.document,
-            self.width - self.margin * 2, self.height - self.margin * 2,
-            multiline=True,
-            batch=self.batch)
+        self.layout = pyglet.text.layout.incremental.IncrementalTextLayout(self.document,
+                                                                           self.width - self.margin * 2, self.height - self.margin * 2,
+                                                                           multiline=True,
+                                                                           batch=self.batch)
         self.caret = caret.Caret(self.layout)
         self.push_handlers(self.caret)
