+++ tests/unit/text/test_layout.py
@@ -1,13 +1,15 @@
 import pyglet
+import pyglet.text.layout.incremental
+
 
 def test_incrementallayout_get_position_on_line_before_start_of_text():
     single_line_text = "This is a single line of text."
     document = pyglet.text.document.UnformattedDocument(single_line_text)
     font = document.get_font()
-    layout = pyglet.text.layout.IncrementalTextLayout(document,
-                                                      height = font.ascent - font.descent,
-                                                      width = 200,
-                                                      multiline=False)
+    layout = pyglet.text.layout.incremental.IncrementalTextLayout(document,
+                                                                  height = font.ascent - font.descent,
+                                                                  width = 200,
+                                                                  multiline=False)
     layout.x = 100
     layout.y = 100
